refactor: keep Spell Focusing reusable without sash
