+++ Projects/UOContent/Misc/AOS.cs
@@ -615,8 +615,7 @@ namespace Server
                 list.Add(1060401, prop); // damage increase ~1_val~%
             }
 
-            // The SA event sash emits its fixed Mana/Defense lines in artifact order from AddNameProperties.
-            if (Owner is not SpellFocusingSash && (prop = DefendChance) != 0)
+            if ((prop = DefendChance) != 0)
             {
                 list.Add(1060408, prop); // defense chance increase ~1_val~%
             }
@@ -671,7 +670,7 @@ namespace Server
                 list.Add(1060436, prop); // luck ~1_val~
             }
 
-            if (Owner is not SpellFocusingSash && (prop = BonusMana) != 0)
+            if ((prop = BonusMana) != 0)
             {
                 list.Add(1060439, prop); // mana increase ~1_val~
             }
@@ -792,7 +791,7 @@ namespace Server
                 {
                     var attrs = clothing.Attributes;
 
-                    if (attrs != null && (clothing is not SpellFocusingSash || Core.SA))
+                    if (attrs != null)
                     {
                         value += attrs[attribute];
                     }
@@ -1604,11 +1603,6 @@ namespace Server
 
         public static bool IsBrittle(Item item)
         {
-            if (item is SpellFocusingSash)
-            {
-                return Core.SA;
-            }
-
             if (!Core.HS)
             {
                 return false;
